move logic to sync and backfill phase
