+++ src/crawl/crawlers/index.js
@@ -5,7 +5,6 @@ import bithomp from './bithomp.js'
 import xrpscan from './xrpscan.js'
 import gravatar from './gravatar.js'
 import x from './x.js'
-import mptmetadata from './mptmetadata.js'
 
 export default [
 	{ name: 'domains', start: domains },
@@ -14,6 +13,5 @@ export default [
 	{ name: 'bithomp', start: bithomp },
 	{ name: 'xrpscan', start: xrpscan },
 	{ name: 'gravatar', start: gravatar },
-	{ name: 'x', start: x },
-	{ name: 'mptmetadata', start: mptmetadata },
+	{ name: 'x', start: x }
 ]
